requirements.txt
Django==3.0.7
django-cors-headers==3.4.0
django-filter==2.3.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
Pillow==7.1.2
psycopg2==2.8.5
